Fix for markdown cells (#2739)
The latest update for the notebook cell handler broke cells in markdown.
Also implements notebook scope handler with a one of scope handler the
same way we do for collection item. We can have both language specific
implementations as well as the ide notebook api so this is a more proper
implementation.

## Checklist

- [x] I have added
[tests](https://www.cursorless.org/docs/contributing/test-case-recorder/)
- [/] I have updated the
[docs](https://github.com/cursorless-dev/cursorless/tree/main/docs) and
[cheatsheet](https://github.com/cursorless-dev/cursorless/tree/main/cursorless-talon/src/cheatsheet)
- [/] I have not broken the cheatsheet

---------
