+++ data/fixtures/recorded/languages/markdown/changeCell.yml
@@ -1,0 +1,27 @@
+languageId: markdown
+command:
+  version: 7
+  spokenForm: change cell
+  action:
+    name: clearAndSetSelection
+    target:
+      type: primitive
+      modifiers:
+        - type: containingScope
+          scopeType: {type: notebookCell}
+  usePrePhraseSnapshot: false
+initialState:
+  documentContents: |
+    ```
+    code
+    ```
+  selections:
+    - anchor: {line: 1, character: 0}
+      active: {line: 1, character: 0}
+  marks: {}
+finalState:
+  documentContents: |+
+
+  selections:
+    - anchor: {line: 0, character: 0}
+      active: {line: 0, character: 0}
